feat: Add Mine and Resource entities
This commit adds the Mine and Resource entities to the codebase. The Mine entity represents a mining operation with properties such as id, name, image, and level. The Resource entity represents a resource with properties like id, name, and icon.

The Mine entity is defined in the 'mines.ts' file, while the Resource entity is defined in the 'resources.ts' file.

These entities will be used to populate the mines and resources data in the application.

Recent user commits:
- feat: Add LeftNavbar component to display navigation links
- chore: Add header and left navbar components to layout.tsx
- chore: Update npm dependencies and add react-icons package
- chore: Add header component for the website
- feat: Add MineCard component to display individual mines on the mines page
- refactor: Update Home component to render a grid of mines